rename OrganizationCell to Organization/OrganizationPickListsCell
